% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie_glm.R
\name{summary.tweedie_glm}
\alias{summary.tweedie_glm}
\title{Summarize a tweedie GLM}
\usage{
\method{summary}{tweedie_glm}(object, method = "Wald", ...)
}
\arguments{
\item{object}{An S3 object returned by tweedie_glm()}

\item{method}{Method to construct confidence interval. Currently only supports 'Wald', but might support 'LR' in the future.}

\item{...}{Other arguments}
}
\description{
This function summarizes a tweedie GLM. By default, it outputs the estimated coefficients and their standard error.
For inference purpose, it also outputs 95% confidence intervals (Wald's test).
}
