% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie_glm.R
\name{tweedie_glm}
\alias{tweedie_glm}
\title{Fits a tweedie GLM model with log link.}
\usage{
tweedie_glm(formula, df, p = NULL, phi = NULL, prior_weights = rep(1,
  nrow(df)), offset = NULL)
}
\arguments{
\item{formula}{Formula of the model}

\item{df}{A data frame. Each column should be either numeric or factor. The response column should be non-negative and numeric}

\item{p}{Power parameter of Tweedie distribution. It is supposed to be in [1, 2], but for numerical reason, it has to be in [1.1, 1.9]}

\item{phi}{Dispersion parameter. It does not affect MLE of coefficients, but is useful in constructing confidence intervals.}

\item{prior_weights}{The weights that appear in the score equation. For example, if the response is loss cost, then the prior_weights
should be exposures. If the response is total loss, then the prior_weights should be $exposure^{p-1}$}

\item{offset}{A numeric vector}
}
\description{
This function fits a tweedie GLM model with log link, power parameter p, dispersion parameter phi. User can also
specify offset and weights. If power parameter p is not specified, it will be estimated before solving for coefficients using MLE.
Parameter phi may be ignored since it does not affect estimation of coefficients. However, if it is provided, the provided
value will be used when constructing confidence interval.
}
