% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie_glmnet.R
\name{tweedie_glmnet}
\alias{tweedie_glmnet}
\title{Tweedie elastic net}
\usage{
tweedie_glmnet(formula, df, p = NULL, phi = NULL,
  prior_weights = rep(1, nrow(df)), offset = NULL, lambda = NULL,
  alpha = NULL)
}
\arguments{
\item{formula}{a formula}

\item{df}{a data frame}

\item{p}{power}

\item{phi}{dispersion}

\item{prior_weights}{weights appear in score equation}

\item{offset}{offset}

\item{lambda}{penalty}

\item{alpha}{penalty of mixture}
}
\description{
Tweedie elastic net
}
